Test or Practice
I noticed that it wasn't really a flashcard program anymore and was more of a testing program. So, I added a feature to select test or practice at the start. The test option runs by allowing the user to type in answers. The Practice option runs by allowing the user to think of the answer and simply press enter to reveal it.